refactor enter_exit
