+++ virtuoso_autonomy/virtuoso_autonomy/robotx/enter_and_exit/loop_point.py
@@ -1,6 +1,5 @@
 from typing import Tuple
 
-
 class LoopPoint:
 
     def __init__(self, change:str, point:Tuple[float, float], dist:float):
